+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.7.0-rc.9
 ## mm/dd/2020
 
+1. [](#new)
+    * Added `Route::getBase()` method
 1. [](#bugfix)
     * Fixed flex objects with integer keys not working [#2863](https://github.com/getgrav/grav/issues/2863)
     * Fixed user avatar creation for new `Flex Users` when using folder storage